+++ src/pages/Allocation/index.tsx
@@ -1,18 +1,14 @@
 import styled from 'styled-components';
 
-import { Button, Text } from 'components/atoms';
 import { Layout, StrategyTitle } from 'components/organism';
 import AllocSetting from './components/AllocSetting';
 import AssetList from './components/AssetList';
 import AllocDateSelect from './components/AllocDateSelect';
+import BackTest from './components/BackTest';
 
-import { flex, theme } from 'styles';
+import { flex } from 'styles';
 
 function Allocation() {
-  const backTest = () => {
-    // code
-  };
-
   return (
     <Layout>
       <Container>
@@ -22,11 +18,7 @@ function Allocation() {
           <AssetList />
           <AllocDateSelect />
         </Wrapper>
-        <BackTestBtnWrap>
-          <BackTestBtn onClick={backTest}>
-            <Text.Medium color={theme.black}>백테스트</Text.Medium>
-          </BackTestBtn>
-        </BackTestBtnWrap>
+        <BackTest />
       </Container>
     </Layout>
   );
@@ -44,14 +36,4 @@ const Wrapper = styled.div`
   ${flex('', '', 'column')};
   width: 640px;
   row-gap: 60px;
-`;
-
-const BackTestBtnWrap = styled.div`
-  text-align: center;
-`;
-
-const BackTestBtn = styled(Button)`
-  width: 190px;
-  padding: 16px 0;
-  background: linear-gradient(to right, rgb(236, 97, 38), rgb(236, 38, 38));
 `;
